+++ wm/sampledata/utils.py
@@ -26,12 +26,15 @@ from wm.sampledata.images import getImage
 from wm.sampledata.images import getFlickrImage
 from wm.sampledata.images import getRandomImage
 from wm.sampledata.images import getRandomFlickrImage
+from wm.sampledata.images import get_placeholder_image
 
 # Silence flake8
 assert getImage
 assert getFlickrImage
 assert getRandomImage
 assert getRandomFlickrImage
+assert get_placeholder_image
+
 
 IPSUM_LINE = "Lorem ipsum mel augue antiopam te. Invidunt constituto accommodare ius cu. Et cum solum liber doming, mel eu quem modus, sea probo putant ex."  # noqa